refactor: moves helper methods to the helpers mod but moves variants function to arg_enum mod where its used
